Confirmando se realmente quer deletar um usuário

+++ src/components/employees/employees.component.ts
@@ -67,6 +67,9 @@ export class EmployeesComponent implements OnInit {
   }
 
   deleteUser(employeeId: string) {
+    if (!confirm('Tem certeza que deseja deletar este usuário? Essa ação é irreversível!')) {
+      return;
+    }
     this.httpClient.delete(`${environment.host}:${environment.port}/${environment.deletePermanUser}/${employeeId}`)
       .subscribe(() => {
         this.loadEmployees();
